+++ .github/workflows/generate_profile_workflow.yml
@@ -29,10 +29,12 @@ jobs:
         uses: actions/setup-python@v2
         with:
           python-version: 3.8 #install the python needed
+
       - name: Install Python dependencies
         uses: py-actions/py-dependency-install@v3
         with:
           path: "./.github/workflows/requirements.txt"
+
       - name: Checkout the DDE repo
         env:
           GITHUB_TOKEN: ${{ secrets.Sahar_Workflows_Token }}
@@ -57,6 +59,18 @@ jobs:
           git pull
           git checkout -b $today
           cd ..
+
+      - name: Tests
+        run: |
+          for /f "skip=1" %%x in ('wmic os get localdatetime') do if not defined MyDate set MyDate=%%x
+          set today=%MyDate:~0,4%-%MyDate:~4,2%-%MyDate:~6,2%
+          echo $today
+          for /f %%x in ('wmic path win32_localtime get /format:list ^| findstr "="') do set %%x
+          set today=%Year%-%Month%-%Day%
+          echo $today
+          for /f %%x in ('wmic path win32_utctime get /format:list ^| findstr "="') do set %%x
+          set today=%Year%-%Month%-%Day%
+          echo $today
 
       - name: Execute the Python Script
         run: |
